Update Aspectator and remove redundant casts from test cases
Also add suddenly missed line directives for a couple of new test cases.

+++ tests/output/c-backend/skip-save-expr.c
@@ -1,4 +1,4 @@
 void func(long unsigned int arg)
 {
-  ( void ) ( long unsigned int ) ( __builtin_constant_p ( arg ) == 0 ? -1 : 1 );
+  ( void ) ( __builtin_constant_p ( arg ) == 0 ? -1 : 1 );
 }
